requirements.txt
google-generativeai
Pillow
streamlit
python-dotenv
